BAI-738 cypress e2e tests now separated into bailo and beta. beta.cjs file created and package.json frontend updated. Now need to update github actions to 2 actions for cypress.
