+++ backend/config/beta.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  experimental: {
+    v2: true,
+  },
+}
