+++ frontend/cypress/e2e/bailo/delete.cy.ts
@@ -1,4 +1,4 @@
-import convertNameToUrlFormat from '../utils/convertNameToUrlFormat'
+import convertNameToUrlFormat from '../../utils/convertNameToUrlFormat'
 
 let modelUrl = ''
