+++ frontend/cypress/e2e/bailo/model.cy.ts
@@ -1,5 +1,5 @@
-import convertNameToUrlFormat from '../utils/convertNameToUrlFormat'
-import getUuidFromUrl from '../utils/getUuidFromUrl'
+import convertNameToUrlFormat from '../../utils/convertNameToUrlFormat'
+import getUuidFromUrl from '../../utils/getUuidFromUrl'
 
 const BASE_URL = Cypress.config('baseUrl')
