+++ frontend/cypress/e2e/bailo/pages/docs.cy.ts
