+++ frontend/cypress/e2e/bailo/pages/index.cy.ts
