+++ frontend/cypress/e2e/bailo/pages/upload.cy.ts
