+++ frontend/cypress/e2e/beta/beta-model.cy.ts
